chore(next): ssr api view
