+++ packages/ui/src/elements/CopyToClipboard/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
