+++ packages/ui/src/exports/forms.ts
@@ -14,5 +14,8 @@ export type { OnChange } from '../forms/field-types/RadioGroup/types'
 export { default as RadioGroupInput } from '../forms/field-types/RadioGroup/Input'
 export { default as Label } from '../forms/Label'
 export { default as Submit } from '../forms/Submit'
+export { default as Checkbox } from '../forms/field-types/Checkbox'
 export { CheckboxInput } from '../forms/field-types/Checkbox/Input'
+export { default as Select } from '../forms/field-types/Select'
 export { default as SelectInput } from '../forms/field-types/Select/Input'
+export { default as Number } from '../forms/field-types/Number'
