+++ packages/ui/src/forms/field-types/Number/types.ts
@@ -1,5 +1,4 @@
 import type { NumberField } from 'payload/types'
+import { FormFieldBase } from '../Text/types'
 
-export type Props = Omit<NumberField, 'type'> & {
-  path?: string
-}
+export type Props = FormFieldBase & Omit<NumberField, 'type'>
